requirements.txt
PyYAML==6.0.2
websocket-client==1.8.0